Fix "int to double" compiler warnings in IpcIoClaimMemoryNeedsRr::run().

+++ src/DiskIO/IpcIo/IpcIoFile.cc
@@ -871,7 +871,8 @@ IpcIoClaimMemoryNeedsRr::run(const RunnerRegistry &)
     // number of queue slots, we add a fudge factor to that to account
     // for corner cases where I/O pages are created before queue
     // limits are checked or destroyed long after the I/O is dequeued
-    Ipc::Mem::NotePageNeed(Ipc::Mem::PageId::ioPage, itemsCount * 1.1);
+    Ipc::Mem::NotePageNeed(Ipc::Mem::PageId::ioPage,
+                           static_cast<int>(itemsCount * 1.1));
 }
